+++ prosessering/pom.xml
@@ -135,6 +135,12 @@
             <artifactId>jackson-datatype-jsr310</artifactId>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>org.jetbrains.kotlinx</groupId>
+            <artifactId>kotlinx-coroutines-core</artifactId>
+            <version>1.3.7</version>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
 
     <build>
