+++ prosessering/src/main/java/no/nav/familie/prosessering/domene/Status.kt
@@ -1,5 +1,8 @@
 package no.nav.familie.prosessering.domene
 
 enum class Status {
-    FERDIG, FEILET, PLUKKET, BEHANDLER, KLAR_TIL_PLUKK, UBEHANDLET, AVVIKSHÅNDTERT, MANUELL_OPPFØLGING
+    FERDIG, FEILET, PLUKKET, BEHANDLER, KLAR_TIL_PLUKK, UBEHANDLET, AVVIKSHÅNDTERT, MANUELL_OPPFØLGING;
+
+
+    fun kanPlukkes() = listOf(KLAR_TIL_PLUKK, UBEHANDLET).contains(this)
 }
