+++ prosessering/src/test/java/no/nav/familie/prosessering/config/ProsesseringConfig.kt
@@ -7,7 +7,7 @@ import org.springframework.scheduling.annotation.EnableAsync
 import org.springframework.scheduling.annotation.EnableScheduling
 import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor
 
-@Configuration @EnableAsync @EnableScheduling
+@Configuration @EnableAsync
 class ProsesseringConfig {
 
     @Bean(name = ["taskExecutor"])
